Fix compiling error in x86 build
And fix also #13811

+++ PowerEditor/src/NppCommands.cpp
@@ -3311,7 +3311,7 @@ void Notepad_plus::command(int id)
 						default:
 							return;
 					}
-					for (size_t i = 0; i < id; i++)
+					for (int i = 0; i < id; i++)
 						wsprintf(hashStr + i * 2, TEXT("%02x"), hash[i]);
 
 					str2Clipboard(hashStr, _pPublicInterface->getHSelf());
